Master Merge

Closes #100 and #103

See merge request cse2000-software-project/2020-2021-q4/cluster-06/water-infrastructure/dhalsim!58

+++ dhalsim/python2/automatic_plant.py
@@ -25,7 +25,8 @@ class PlantControl(NodeControl):
         This function stops the physical process (child of this process).
         """
         self.logger.debug("Stopping plant.")
-        self.simulation_process.send_signal(signal.SIGINT)
+        if self.simulation_process.poll() is None:
+            self.simulation_process.send_signal(signal.SIGINT)
         self.simulation_process.wait()
         if self.simulation_process.poll() is None:
             self.simulation_process.terminate()
@@ -45,6 +46,8 @@ class PlantControl(NodeControl):
 
         while self.simulation_process.poll() is None:
             pass
+
+        self.terminate()
 
 
 def is_valid_file(parser_instance, arg):
